// src/app/shared/models/planilla-mensual-response.models.ts
export interface PlanillaMensualResponse {
idPlanillaMensual?: number;
  anio: number;
  mes: number;
  trabajador?: any;
  situacion?: any;
  cargo?: any;
  apellido?: string;
  nombre?: string;
  sistemaPension?: any;
  estadoCivil?: any;
  hijos?: number;
  fechaIngreso?: string;
  sueldoBasico?: number;
  porcHoraExtra1?: number;
  porcHoraExtra2?: number;
  porcDescansoTrab?: number;
  porcFeriadoTrab?: number;
  PorcAsigFamiliar?: number;
  nHorasNormal?: number;
  nHorasExtra1?: number;
  nHorasExtra2?: number;
  nDiasTrab?: number;
  nDiasDescansos?: number;
  nFeriadosTrab?: number;
  nDescansosTrab?: number;
  nDiasInasistencias?: number;
  haberBasico?: number;
  valesEmpleado?: number;
  vHorasExtra1?: number;
  vHorasExtra2?: number;
  vAsigFamiliar?: number;
  vDescansoTrab?: number;
  vFeriadoTrab?: number;
  bonificacionCargo?: number;
  bonificacionMovilidad?: number;
  canastaNavidad?: number;
  escolaridad?: number;
  diaTrabajador?: number;
  totalIngreso?: number;
  renta5ta?: number;
  descuentoJud1?: number;
  descuentoJud2?: number;
  descuentoJud3?: number;
  otrosAdelantos?: number;
  adelantoCaja?: number;
  adelantoQuincena?: number;
  adelantoVac?: number;
  adelantoGratificacion?: number;
  adelantoLiquidacion?: number;
  adelantoCTS?: number;
  porcAporte?: number;
  aporte?: number;
  porcComision?: number;
  comision?: number;
  porcPrima?: number;
  prima?: number;
  otdSeg?: number;
  otdPacifico?: number;
  idBanco1?: number;
  prestamo1?: number;
  tardanza?: number;
  totalDescuento?: number;
  porcEsSalud?: number;
  esSalud?: number;
  accidenteTrab?: number;
  senati?: number;
  seguroVidaLey?: number;
  totalNeto?: number;
  totalNetoCad?: string;
  totalNetoBoleta?: number;
  TotalNetoBoletaCad?: string;
  
  // Campos específicos de búsqueda por documento
  vhorasExtra1?: number;
  vhorasExtra2?: number;
  vasigFamiliar?: number;
  vferiadoTrab?: number;
  ndiasTrab?: number;
  nhorasNormal?: number;
}